better socket interface with singleton pattern and better events
